rename directory to be more descriptive
Also start adding in provider/context to the individual steps
