+++ components/Workflows/NewSourceFlow/Steps/SourceTypeSelection.js
@@ -1,6 +1,5 @@
 import Image from "next/image"
 
-import Button from "@components/Button"
 import { getLogoForSourceType } from "@hooks/useSourceLogos"
 
 export default function SourceTypeSelection({
@@ -37,10 +36,6 @@ export default function SourceTypeSelection({
 
   return (
     <div className="flex h-full flex-col overflow-hidden">
-      <div className="flex items-center justify-between border-b border-neutral-100 px-6 py-4">
-        <h2 className="text-xl font-semibold leading-none">Select a Source Type</h2>
-        <Button onClick={onBack}>Back</Button>
-      </div>
       <div className="h-full overflow-y-auto p-6">
         <div className="grid grid-cols-2 gap-4">
           {filteredSourceTypes.map((sourceType) => {
